Changing Floats to Doubles.

+++ TrafficDataContracts/src/main/java/fi/ahto/example/traffic/data/contracts/internal/VehicleActivity.java
@@ -205,19 +205,19 @@ public class VehicleActivity implements Serializable {
         this.source = source;
     }
 
-    public Float getSpeed() {
+    public Double getSpeed() {
         return speed;
     }
 
-    public void setSpeed(Float speed) {
+    public void setSpeed(Double speed) {
         this.speed = speed;
     }
 
-    public Float getBearing() {
+    public Double getBearing() {
         return bearing;
     }
 
-    public void setBearing(Float bearing) {
+    public void setBearing(Double bearing) {
         this.bearing = bearing;
     }
 
@@ -317,19 +317,19 @@ public class VehicleActivity implements Serializable {
         this.direction = direction;
     }
 
-    public Float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
-    public Float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
@@ -348,9 +348,9 @@ public class VehicleActivity implements Serializable {
     @JsonProperty("Direction")
     private String direction;
     @JsonProperty("Longitude")
-    private Float longitude;
+    private Double longitude;
     @JsonProperty("Latitude")
-    private Float latitude;
+    private Double latitude;
     @JsonProperty("Delay")
     private Integer delay;
     @JsonProperty("InternalLineId")
@@ -374,9 +374,9 @@ public class VehicleActivity implements Serializable {
     @JsonProperty("LineHasChanged")
     private boolean lineHasChanged;
     @JsonProperty("Bearing")
-    private Float bearing;
+    private Double bearing;
     @JsonProperty("Speed")
-    private Float speed;
+    private Double speed;
     @JsonProperty("Source")
     private String source;
     @JsonProperty("NextStopSequence")
